+++ packages/opentelemetry-core/src/common/time.ts
@@ -19,18 +19,12 @@ import { otperformance as performance } from '../platform';
 import { TimeOriginLegacy } from './types';
 
 const NANOSECOND_DIGITS = 9;
+const NANOSECOND_DIGITS_IN_MILLIS = 6;
+const MILLISECONDS_TO_NANOSECONDS = Math.pow(10, NANOSECOND_DIGITS_IN_MILLIS);
 const SECOND_TO_NANOSECONDS = Math.pow(10, NANOSECOND_DIGITS);
 
 /**
- * Converts a number to HrTime, HrTime = [number, number].
- * The first number is UNIX Epoch time in seconds since 00:00:00 UTC on 1 January 1970.
- * The second number represents the partial second elapsed since Unix Epoch time represented by first number in nanoseconds.
- * For example, 2021-01-01T12:30:10.150Z in UNIX Epoch time in milliseconds is represented as 1609504210150.
- * numberToHrtime calculates the first number by converting and truncating the Epoch time in milliseconds to seconds:
- * HrTime[0] = Math.trunc(1609504210150 / 1000) = 1609504210.
- * numberToHrtime calculates the second number by converting the digits after the decimal point of the subtraction, (1609504210150 / 1000) - HrTime[0], to nanoseconds:
- * HrTime[1] = Number((1609504210.150 - HrTime[0]).toFixed(9)) * SECOND_TO_NANOSECONDS = 150000000.
- * This is represented in HrTime format as [1609504210, 150000000].
+ * Converts a number of milliseconds from epoch to HrTime([seconds, remainder in nanoseconds]).
  * @param epochMillis
  */
 function numberToHrtime(epochMillis: number): api.HrTime {
@@ -38,9 +32,7 @@ function numberToHrtime(epochMillis: number): api.HrTime {
   // Decimals only.
   const seconds = Math.trunc(epochSeconds);
   // Round sub-nanosecond accuracy to nanosecond.
-  const nanos =
-    Number((epochSeconds - seconds).toFixed(NANOSECOND_DIGITS)) *
-    SECOND_TO_NANOSECONDS;
+  const nanos = Math.round((epochMillis % 1000) * MILLISECONDS_TO_NANOSECONDS);
   return [seconds, nanos];
 }